+++ include/celestialobject.h
@@ -29,7 +29,7 @@ public:
 	bool getCollision(sf::Vector2f& point); // Collision check for points
 	bool getCollision(sf::CircleShape&); // Collision check for CircleShape
 	void onCollision(CelestialObject& other);
-	float getAngle(sf::CircleShape&); // Returns angle compared to a object in radians
+	float getAngle(CelestialObject&); // Returns angle compared to a object in radians
 };
 
 #endif