+++ src/navball.cpp
@@ -1,6 +1,9 @@
 #include "graphics.h"
 #include "navball.h"
 
+#include <cmath>
+#include <iostream>
+
 Navball::Navball(Spaceship* spaceship) {
     this->spaceship = spaceship;
     sf::RenderWindow* window = getWindow();
@@ -26,6 +29,15 @@ Navball::Navball(Spaceship* spaceship) {
     this->directionPointer.setPosition(pos.x, pos.y);
     this->directionPointer.setOrigin(2,0);
 
+    this->movementPointer.setSize(sf::Vector2f(4,radius+10));
+    this->movementPointer.setPosition(pos.x, pos.y);
+    this->movementPointer.setOrigin(2,0);
+    this->movementPointer.setFillColor(sf::Color::Yellow);
+
+    this->destinationPointer.setSize(sf::Vector2f(4,radius+10));
+    this->destinationPointer.setPosition(pos.x, pos.y);
+    this->destinationPointer.setOrigin(2,0);
+
     this->distanceText = sf::Text("",getFont(), 30);
     this->speedText = sf::Text("", getFont(), 30);
     this->distanceText.setPosition(pos.x+radius, pos.y);
@@ -33,10 +45,28 @@ Navball::Navball(Spaceship* spaceship) {
 }
 
 void Navball::draw(sf::RenderWindow& window, CelestialObject* selectedObject){
-    this->directionPointer.setRotation((this->spaceship->getAngle()/3.14)*180+90);
+    float directionAngle = (this->spaceship->getMovementAngle()/3.14)*180+90;
+    this->directionPointer.setRotation(directionAngle);
+
+    float movementAngle = std::atan2(this->spaceship->getSpeed().x,
+                                     this->spaceship->getSpeed().y);
+    this->movementPointer.setRotation((movementAngle/3.14)*180+180);
+
+    if (selectedObject != nullptr){
+        float destinationAngle = this->spaceship->getAngle(*selectedObject);
+        this->destinationPointer.setRotation((-destinationAngle/3.14)*180+90);
+        std::cout << destinationAngle*180+90 << std::endl;
+        this->destinationPointer.setFillColor(sf::Color::Blue);
+    }
+    else {
+        this->destinationPointer.setFillColor(sf::Color(125,125,125));
+    }
+
 
     window.draw(this->frame);
     window.draw(this->directionPointer);
+    window.draw(this->movementPointer);
+    window.draw(this->destinationPointer);
     window.draw(this->background);
 
     if (selectedObject != nullptr){